\name{plot.resample.sprinter}
\alias{plot.resample.sprinter}
\title{Plots coefficients generated by \code{resample.sprinter}}
\description{
Plots for each variable which is selected in \code{resample.sprinter} the mean coefficient over all subsamples against its single coefficients.
}
\usage{
\method{plot}{resample.sprinter}(x, optional.only = FALSE, threshold.vif = 0,\ldots)
}
\arguments{
\item{x}{resample.sprinter object from a \code{\link{resample.sprinter}} call.}
\item{optional.only}{logical value. If true, only the coefficients of the optional variables are plotted. If false, the coefficients of the mandatory variables are plotted as well.}
\item{threshold.vif}{value. Only the coefficients with inclusion frequencies larger than \code{threshold.vif} are plotted.}
\item{\dots}{additional arguments.}
}
\value{
No value is returned, but a plot is generated.
}
\author{
Isabell Hoffmann \email{isabell.hoffmann@uni-mainz.de}
}
