\name{summary.sprinter}
\alias{summary.sprinter}
\title{Generates a summary for objects of class \code{sprinter}}
\description{
Shows the main candidates with coefficients of the main effects model together with the interaction candidates with largest inclusion frequencies and prints the final model.
}
\usage{
\method{summary}{sprinter}(object, \ldots)
}
\arguments{
\item{object}{Object of class \code{\link{sprinter}}.}
\item{\ldots}{additional arguments.}
}


\author{
Isabell Hoffmann \email{isabell.hoffmann@uni-mainz.de}
}

